@@ -6,7 +6,8 @@ description: Demonstrates how to use the video_player_win plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: '>=2.17.0 < 4.0.0'
+  sdk: '>=2.17.0'
+  flutter: ">=2.20.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -19,12 +20,11 @@ dependencies:
     sdk: flutter
 
   video_player_win:
-    # When depending on this package from a real application you should use:
-    #   video_player_win: ^x.y.z
-    # See https://dart.dev/tools/pub/dependencies#version-constraints
-    # The example app is bundled with the plugin so we use a path dependency on
-    # the parent directory to use the current plugin's version.
     path: ..
+
+  video_player_platform_interface: ^6.2.1
+  video_player: ^2.9.1
+  shared_preferences: ^2.2.2
 
 
   # The following adds the Cupertino Icons font to your application.
